create acquisition_bmotion.py and move in run_acquisition_bmotion() from multi_scope_acquisition.py
